Don't use row length checks or hardcoded column numbers with SQLite
Hardcoded column numbers are unnecessary and magical,
and the row length checks were brittle and unnecessary.

This PR introduces a TableRow abstraction that provides methods for
reading column values that give descriptive error messages if a
required column is missing or doesn't have a value in a particular row.
It then provides a small VisitTableRows function that serves as a richer
equivalent to the VisitTableRecords method from go-sqlite.
The ReadCookies methods for chrome and firefox are rewritten in terms of
these abstractions.