Fix issues introduced in the previous commit
The previous commit failed to properly signal errors in Int and Byte
values. This masked two additional issues:
- These handlers didn't handle all integer types they could encounter.
- Older chrome databases use alternate names for the is_secure and
  is_httponly headers.

The previous commit also failed to handle uint64 values when we
expect an int64. This is now handled, which a check for values
that would result in an incorrect interpretation after conversion.